% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_14_tutorial.R
\name{chapter_14_tutorial}
\alias{chapter_14_tutorial}
\title{Chapter 14 Tutorial}
\usage{
chapter_14_tutorial()
}
\value{

}
\description{
Runs the `learnr` code for chapter 14: Reproducibility and Coding Best Practices
}
