% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_4_tutorial.R
\name{chapter_4_tutorial}
\alias{chapter_4_tutorial}
\title{Chapter 4 Tutorial}
\usage{
chapter_4_tutorial()
}
\value{

}
\description{
Runs the `learnr` code for chapter 4: Intro to plotting with `ggplot2`
}
