% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_7_tutorial.R
\name{chapter_7_tutorial}
\alias{chapter_7_tutorial}
\title{Chapter 7 Tutorial}
\usage{
chapter_7_tutorial()
}
\value{

}
\description{
Runs the `learnr` code for chapter 7: Tibble Joins: A Tale of Two Tibbles
}
