% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter_8_tutorial.R
\name{chapter_8_tutorial}
\alias{chapter_8_tutorial}
\title{Chapter 8 Tutorial}
\usage{
chapter_8_tutorial()
}
\value{

}
\description{
Runs the `learnr` code for chapter 8: Tibble Joins: File input/output (I/O)
}
