% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_homepage.R
\name{open_homepage}
\alias{open_homepage}
\title{Goes to the Data Science and Programming class homepage}
\usage{
open_homepage()
}
\value{

}
\description{
Goes to the Data Science and Programming class homepage
}
