% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_tutorial_best_practices.R
\name{tutorial_best_practices}
\alias{tutorial_best_practices}
\title{Coding Best Practices Tutorial}
\usage{
tutorial_best_practices()
}
\value{

}
\description{
Runs the `learnr` code for Chapter 14: Reproducibility and Coding Best Practices
}
