% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_tutorial_file_io.R
\name{tutorial_file_io}
\alias{tutorial_file_io}
\title{File Input/Output Tutorial}
\usage{
tutorial_file_io()
}
\value{

}
\description{
Runs the `learnr` code for Chapter 7: File input/output (I/O)
}
